# Use labels Bytes() rather than String() for map key