Signed-off-by: Ben Ye <benye@amazon.com>

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with master
-->

**What this PR does**:

Use `string(labels.Bytes())` instead of `labels.String()` for the deduplication map key. Since we don't have to print the map key, we just use the key to find the samples and as long as it remains the same sorting order, we can safely replace the method.

Benchmark 

```
name    old time/op    new time/op    delta
Label-10    9.60µs ± 0%    1.60µs ± 0%   ~     (p=1.000 n=1+1)

name    old alloc/op   new alloc/op   delta
Label-10    7.79kB ± 0%    3.71kB ± 0%   ~     (p=1.000 n=1+1)

name    old allocs/op  new allocs/op  delta
Label-10       111 ± 0%         2 ± 0%   ~     (p=1.000 n=1+1)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`